feat(test): 重整osx目录测试文件结构
将9个分散的测试文件合并为1个file_test.go文件，遵循测试文件命名规范：
- 删除: descriptor_test.go, direct_stat_test.go, edge_cases_test.go等8个文件
- 合并到: file_test.go（与源文件file.go对应）
- 保持97.7%测试覆盖率
- 保留所有27个测试函数和107个子测试
