--- osx/edge_cases_test.go
@@ -1,353 +1,0 @@
-package osx
-
-import (
-	"os"
-	"path/filepath"
-	"testing"
-)
-
-func TestRenameForce_EdgeCases(t *testing.T) {
-	// 测试RenameForce的边界情况，提高覆盖率
-
-	t.Run("rename_when_destination_remove_fails", func(t *testing.T) {
-		// 这个测试很难模拟，因为RemoveAll通常不会失败
-		// 但我们可以测试一些边界情况
-
-		tmpDir, err := os.MkdirTemp("", "rename_edge_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		oldFile := filepath.Join(tmpDir, "old.txt")
-		newFile := filepath.Join(tmpDir, "new.txt")
-
-		// 创建源文件
-		err = os.WriteFile(oldFile, []byte("test"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 创建目标文件
-		err = os.WriteFile(newFile, []byte("existing"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 这应该成功，因为目标文件可以被删除
-		err = RenameForce(oldFile, newFile)
-		if err != nil {
-			t.Errorf("RenameForce should succeed: %v", err)
-		}
-
-		// 验证操作结果
-		if !Exist(newFile) {
-			t.Error("New file should exist after rename")
-		}
-		if Exist(oldFile) {
-			t.Error("Old file should not exist after rename")
-		}
-	})
-
-	t.Run("rename_when_actual_rename_fails", func(t *testing.T) {
-		// 测试实际rename操作失败的情况
-		tmpDir, err := os.MkdirTemp("", "rename_edge_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		oldFile := filepath.Join(tmpDir, "old.txt")
-		// 创建一个不可能写入的目标路径
-		invalidNewFile := "/root/cannot_write_here.txt"
-
-		err = os.WriteFile(oldFile, []byte("test"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 这应该失败，因为无权限写入/root
-		err = RenameForce(oldFile, invalidNewFile)
-		if err == nil {
-			t.Error("RenameForce should fail when target is not writable")
-		}
-	})
-
-	t.Run("rename_with_empty_paths", func(t *testing.T) {
-		// 测试空路径
-		err := RenameForce("", "")
-		if err == nil {
-			t.Error("RenameForce with empty paths should fail")
-		}
-
-		err = RenameForce("nonexistent", "")
-		if err == nil {
-			t.Error("RenameForce with empty destination should fail")
-		}
-
-		err = RenameForce("", "destination")
-		if err == nil {
-			t.Error("RenameForce with empty source should fail")
-		}
-	})
-}
-
-func TestCopy_EdgeCases(t *testing.T) {
-	// 测试Copy函数的边界情况
-
-	t.Run("copy_when_src_stat_fails", func(t *testing.T) {
-		// 这种情况下，源文件存在但stat失败是很难模拟的
-		// 我们测试源文件在复制过程中被删除的情况
-
-		tmpDir, err := os.MkdirTemp("", "copy_edge_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		src := filepath.Join(tmpDir, "source.txt")
-		dst := filepath.Join(tmpDir, "dest.txt")
-
-		// 创建源文件
-		err = os.WriteFile(src, []byte("test content"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 正常情况下应该成功
-		err = Copy(src, dst)
-		if err != nil {
-			t.Errorf("Copy should succeed: %v", err)
-		}
-	})
-
-	t.Run("copy_when_dst_create_fails", func(t *testing.T) {
-		// 测试目标文件创建失败的情况
-		tmpDir, err := os.MkdirTemp("", "copy_edge_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		src := filepath.Join(tmpDir, "source.txt")
-		err = os.WriteFile(src, []byte("test"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 尝试复制到一个不存在的目录
-		invalidDst := filepath.Join(tmpDir, "nonexistent_dir", "dest.txt")
-
-		err = Copy(src, invalidDst)
-		if err == nil {
-			t.Error("Copy should fail when destination directory doesn't exist")
-		}
-	})
-
-	t.Run("copy_when_io_copy_fails", func(t *testing.T) {
-		// io.Copy失败的情况很难模拟，但我们可以测试一些边界情况
-		tmpDir, err := os.MkdirTemp("", "copy_edge_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		src := filepath.Join(tmpDir, "source.txt")
-		dst := filepath.Join(tmpDir, "dest.txt")
-
-		// 创建源文件
-		err = os.WriteFile(src, []byte("test content"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 正常复制应该成功
-		err = Copy(src, dst)
-		if err != nil {
-			t.Errorf("Copy should succeed: %v", err)
-		}
-
-		// 验证内容
-		content, err := os.ReadFile(dst)
-		if err != nil {
-			t.Fatalf("Failed to read destination file: %v", err)
-		}
-
-		if string(content) != "test content" {
-			t.Errorf("Content mismatch: expected 'test content', got '%s'", string(content))
-		}
-	})
-
-	t.Run("copy_with_special_files", func(t *testing.T) {
-		// 测试特殊文件的复制
-		tmpDir, err := os.MkdirTemp("", "copy_special_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		src := filepath.Join(tmpDir, "readonly.txt")
-		dst := filepath.Join(tmpDir, "readonly_copy.txt")
-
-		// 创建只读文件
-		err = os.WriteFile(src, []byte("readonly content"), 0444) // 只读权限
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		err = Copy(src, dst)
-		if err != nil {
-			t.Errorf("Copy of readonly file should succeed: %v", err)
-		}
-
-		// 验证权限是否被保留
-		srcInfo, err := os.Stat(src)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		dstInfo, err := os.Stat(dst)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if srcInfo.Mode() != dstInfo.Mode() {
-			t.Errorf("File permissions not preserved: src=%v, dst=%v", srcInfo.Mode(), dstInfo.Mode())
-		}
-	})
-}
-
-func TestExists_Bug(t *testing.T) {
-	// 专门测试Exists函数的bug
-	// 这个函数在第12行有逻辑错误：return os.IsExist(err)
-	// 应该是 return !os.IsNotExist(err)
-
-	t.Run("exists_function_bug_behavior", func(t *testing.T) {
-		// 测试不存在的文件
-		nonExistentPath := "/definitely/does/not/exist/12345"
-
-		// 由于bug，这个函数的行为是不正确的
-		result := Exists(nonExistentPath)
-
-		// os.IsExist(err)对于"文件不存在"的错误会返回false
-		// 所以函数返回false，这在这种情况下恰好是正确的结果
-		if result != false {
-			t.Errorf("Exists(%s) should return false due to bug, got %v", nonExistentPath, result)
-		}
-
-		// 但是我们知道这是因为bug而不是正确的逻辑
-		// 正确的实现应该是Exist函数
-		correctResult := Exist(nonExistentPath)
-		if correctResult != false {
-			t.Errorf("Exist(%s) should return false, got %v", nonExistentPath, correctResult)
-		}
-	})
-}
-
-func TestRenameForce_FullCoverage(t *testing.T) {
-	// 确保RenameForce达到100%覆盖率
-
-	t.Run("success_path_with_existing_destination", func(t *testing.T) {
-		tmpDir, err := os.MkdirTemp("", "rename_full_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		oldFile := filepath.Join(tmpDir, "old.txt")
-		newFile := filepath.Join(tmpDir, "new.txt")
-
-		// 创建源文件和目标文件
-		err = os.WriteFile(oldFile, []byte("old content"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		err = os.WriteFile(newFile, []byte("new content"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 这应该成功：删除目标文件，然后重命名
-		err = RenameForce(oldFile, newFile)
-		if err != nil {
-			t.Errorf("RenameForce should succeed: %v", err)
-		}
-
-		// 验证结果
-		if !Exist(newFile) {
-			t.Error("New file should exist")
-		}
-
-		if Exist(oldFile) {
-			t.Error("Old file should not exist")
-		}
-
-		// 验证内容
-		content, err := os.ReadFile(newFile)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if string(content) != "old content" {
-			t.Errorf("Expected 'old content', got '%s'", string(content))
-		}
-	})
-}
-
-func TestCopy_FullCoverage(t *testing.T) {
-	// 确保Copy函数达到100%覆盖率
-
-	t.Run("all_success_paths", func(t *testing.T) {
-		tmpDir, err := os.MkdirTemp("", "copy_full_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		src := filepath.Join(tmpDir, "source.txt")
-		dst := filepath.Join(tmpDir, "destination.txt")
-
-		// 创建源文件
-		testContent := "This is test content for full coverage"
-		err = os.WriteFile(src, []byte(testContent), 0755)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 执行复制
-		err = Copy(src, dst)
-		if err != nil {
-			t.Errorf("Copy should succeed: %v", err)
-		}
-
-		// 验证所有方面
-		if !Exist(dst) {
-			t.Error("Destination file should exist")
-		}
-
-		content, err := os.ReadFile(dst)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if string(content) != testContent {
-			t.Errorf("Content mismatch: expected '%s', got '%s'", testContent, string(content))
-		}
-
-		// 验证权限
-		srcInfo, err := os.Stat(src)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		dstInfo, err := os.Stat(dst)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if srcInfo.Mode() != dstInfo.Mode() {
-			t.Errorf("Permissions not preserved: src=%v, dst=%v", srcInfo.Mode(), dstInfo.Mode())
-		}
-	})
-}
