--- osx/error_branch_test.go
@@ -1,303 +1,0 @@
-package osx
-
-import (
-	"os"
-	"path/filepath"
-	"testing"
-)
-
-func TestRenameForce_RemoveAllError(t *testing.T) {
-	// 尝试触发os.RemoveAll错误分支（第53-55行）
-
-	t.Run("remove_readonly_directory", func(t *testing.T) {
-		if os.Geteuid() == 0 {
-			t.Skip("Skipping test when running as root")
-		}
-
-		tmpDir, err := os.MkdirTemp("", "rename_remove_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		oldFile := filepath.Join(tmpDir, "old.txt")
-		newDir := filepath.Join(tmpDir, "readonly_dir")
-
-		// 创建源文件
-		err = os.WriteFile(oldFile, []byte("test"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 创建一个包含文件的目录
-		err = os.Mkdir(newDir, 0755)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		subFile := filepath.Join(newDir, "file.txt")
-		err = os.WriteFile(subFile, []byte("content"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 尝试将目录设为只读（这在某些系统上可能不会阻止删除）
-		err = os.Chmod(newDir, 0444)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 尝试重命名到这个目录
-		err = RenameForce(oldFile, newDir)
-
-		// 恢复权限以便清理
-		os.Chmod(newDir, 0755)
-
-		// 在大多数Unix系统上，这个操作实际上会成功
-		// 但我们至少执行了这个代码路径
-		t.Logf("RenameForce result: %v", err)
-	})
-
-	t.Run("remove_permission_denied", func(t *testing.T) {
-		// 在某些情况下可能触发权限错误
-		tmpDir, err := os.MkdirTemp("", "rename_perm_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		oldFile := filepath.Join(tmpDir, "old.txt")
-		newFile := filepath.Join(tmpDir, "new.txt")
-
-		// 创建源文件和目标文件
-		err = os.WriteFile(oldFile, []byte("old content"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		err = os.WriteFile(newFile, []byte("new content"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 在大多数情况下，这会成功
-		err = RenameForce(oldFile, newFile)
-		if err != nil {
-			t.Logf("RenameForce failed (expected in some cases): %v", err)
-		} else {
-			t.Logf("RenameForce succeeded")
-			// 验证操作成功
-			if !Exist(newFile) {
-				t.Error("New file should exist after successful rename")
-			}
-		}
-	})
-}
-
-func TestCopy_StatError(t *testing.T) {
-	// 尝试触发srcFile.Stat()错误分支（第73-75行）
-
-	t.Run("stat_error_after_open", func(t *testing.T) {
-		// 这个错误很难触发，因为如果文件能被打开，通常也能被stat
-		// 但我们可以测试一些边界情况
-
-		tmpDir, err := os.MkdirTemp("", "copy_stat_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		src := filepath.Join(tmpDir, "source.txt")
-		dst := filepath.Join(tmpDir, "dest.txt")
-
-		// 创建源文件
-		err = os.WriteFile(src, []byte("test content"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 正常情况下应该成功
-		err = Copy(src, dst)
-		if err != nil {
-			t.Errorf("Copy should succeed: %v", err)
-		} else {
-			// 验证复制成功
-			content, err := os.ReadFile(dst)
-			if err != nil {
-				t.Fatal(err)
-			}
-			if string(content) != "test content" {
-				t.Errorf("Content mismatch: expected 'test content', got '%s'", string(content))
-			}
-		}
-	})
-}
-
-func TestCopy_IOCopyError(t *testing.T) {
-	// 尝试触发io.Copy错误分支（第84-86行）
-
-	t.Run("disk_space_exhaustion_simulation", func(t *testing.T) {
-		// io.Copy错误很难模拟，但我们可以测试一些情况
-		tmpDir, err := os.MkdirTemp("", "copy_io_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		src := filepath.Join(tmpDir, "source.txt")
-		dst := filepath.Join(tmpDir, "dest.txt")
-
-		// 创建源文件
-		testContent := "This is test content for io.Copy testing"
-		err = os.WriteFile(src, []byte(testContent), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 正常情况下应该成功
-		err = Copy(src, dst)
-		if err != nil {
-			t.Logf("Copy failed (might indicate IO error): %v", err)
-		} else {
-			// 验证复制成功
-			if !Exist(dst) {
-				t.Error("Destination file should exist after successful copy")
-			}
-
-			content, err := os.ReadFile(dst)
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			if string(content) != testContent {
-				t.Errorf("Content mismatch: expected '%s', got '%s'", testContent, string(content))
-			}
-		}
-	})
-
-	t.Run("copy_zero_byte_file", func(t *testing.T) {
-		// 测试零字节文件复制
-		tmpDir, err := os.MkdirTemp("", "copy_zero_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		src := filepath.Join(tmpDir, "empty.txt")
-		dst := filepath.Join(tmpDir, "empty_copy.txt")
-
-		// 创建空文件
-		err = os.WriteFile(src, []byte(""), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		err = Copy(src, dst)
-		if err != nil {
-			t.Errorf("Copy of empty file should succeed: %v", err)
-		} else {
-			// 验证复制的空文件
-			content, err := os.ReadFile(dst)
-			if err != nil {
-				t.Fatal(err)
-			}
-			if len(content) != 0 {
-				t.Errorf("Expected empty file, got %d bytes", len(content))
-			}
-		}
-	})
-}
-
-func TestCopy_OpenFileError(t *testing.T) {
-	// 测试目标文件创建失败的情况
-
-	t.Run("create_in_readonly_dir", func(t *testing.T) {
-		if os.Geteuid() == 0 {
-			t.Skip("Skipping test when running as root")
-		}
-
-		tmpDir, err := os.MkdirTemp("", "copy_readonly_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		src := filepath.Join(tmpDir, "source.txt")
-		readonlyDir := filepath.Join(tmpDir, "readonly")
-		dst := filepath.Join(readonlyDir, "dest.txt")
-
-		// 创建源文件
-		err = os.WriteFile(src, []byte("test"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 创建只读目录
-		err = os.Mkdir(readonlyDir, 0755)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		err = os.Chmod(readonlyDir, 0444) // 只读权限
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 这应该失败，因为无法在只读目录中创建文件
-		err = Copy(src, dst)
-
-		// 恢复权限以便清理
-		os.Chmod(readonlyDir, 0755)
-
-		if err == nil {
-			t.Error("Copy should fail when destination directory is readonly")
-		} else {
-			t.Logf("Copy failed as expected: %v", err)
-		}
-	})
-}
-
-func TestExists_AllBranches(t *testing.T) {
-	// 确保Exists函数的所有分支都被测试
-
-	t.Run("exists_with_permission_error", func(t *testing.T) {
-		// 测试权限错误情况
-		if os.Geteuid() == 0 {
-			t.Skip("Skipping test when running as root")
-		}
-
-		tmpDir, err := os.MkdirTemp("", "exists_perm_test_*")
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer os.RemoveAll(tmpDir)
-
-		// 创建一个文件
-		testFile := filepath.Join(tmpDir, "test.txt")
-		err = os.WriteFile(testFile, []byte("test"), 0644)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 移除目录的执行权限
-		err = os.Chmod(tmpDir, 0000)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// 现在尝试检查文件是否存在
-		result := Exists(testFile)
-
-		// 恢复权限
-		os.Chmod(tmpDir, 0755)
-
-		// 由于权限问题，os.Stat会返回错误
-		// 由于Exists函数的bug，它会调用os.IsExist(err)
-		// 对于权限错误，这通常返回false
-		t.Logf("Exists result for permission-denied file: %v", result)
-
-		// 对比正确的Exist函数
-		correctResult := Exist(testFile)
-		t.Logf("Exist result for same file: %v", correctResult)
-	})
-}
